Update testcases to include devcontainer
`test_it_can_return_the_sail_install_script` is updated
because we're adding an extra placeholder (with a space)
to the default output. I think it's okay to remove the final
quotation mark because it appears we're just testing to see
if the output resembles what we expect.

`test_it_adds_the_devcontainer_upon_request` is self
explanatory.

Composer lock was updated incidentally.